Implement athanor project dependency
In this change, we remove the athanor jar that ideally belongs to
the athanor project, and resolve a versioned distribution jar
that is automatically downloaded with the athanor project
which is now a dependency of this project. An athanor.jar
link in the athanor project directory points to the
versioned jar that we are dependent on my default.